// src/modules/logistics-analyst/components/tabs/info/types.ts
export interface RF1 {
  country: string;
  street: string;
  internal_number: string;
  external_number: string;
  zip_code: string;
  city: string;
  municipality: string;
  colony: string;
  tradeName: string;
  genre: string;
  birhdate: string;
  phone_natural_person: string;
  rcPolicy: boolean;
  effectiveDateStart: string;
  effectiveDateEnd: string;
  direct_employees: string;
  employees_quantity: string;
  annual_income: string;
  annual_income_label: string;
  finalized: boolean;
  addressString1: string;
  addressString2: string;
  phone_legal_person: string;
  add_branch: boolean;
  branches: any;
}